delete user almost working some issues with database

+++ server/participant/participant-controller.js
@@ -19,6 +19,7 @@ router.get(
       return;
     }
     Participant.findAll({
+      attributes: ["id", "email", "is_admin", "is_mod", "moderator_id"],
       where: {
         moderator_id: jwt.user.id,
       },
@@ -34,18 +35,26 @@ router.get(
   }
 );
 
-router.delete(
+router.post(
   "/deleteParticipant",
   passport.authenticate("jwt", { session: false }),
   (req, res) => {
+    console.log(req.body);
     jwt = server.getJwt(req);
+    if (!jwt.user.is_mod) {
+      res.status(401);
+      return;
+    }
     Participant.destroy({
       where: {
-        id: req.body.id,
+        id: req.body,
       },
     })
-      .then((users) => {})
+      .then((users) => {
+        res.status(200).send("successs");
+      })
       .catch((err) => {
+        console.log(err);
         res.status(400).send(err);
       });
   }
